rename to `gkr_proof` mod to `logup_gkr`
