+++ processor/src/trace/virtual_bus/univariate.rs
@@ -1,5 +1,5 @@
 use alloc::vec::Vec;
-use miden_air::gkr_proof::UnivariatePolyCoef;
+use miden_air::logup_gkr::UnivariatePolyCoef;
 use vm_core::FieldElement;
 use winter_prover::math::batch_inversion;
